Added PATH_SRC const for path to images and css and Renamed service to services
