+++ src/ducks/goods.js
@@ -1,7 +1,7 @@
 import { createAction, createReducer } from 'redux-act'
 import axios from 'axios'
 import _ from 'lodash'
-import { API } from '../service'
+import { API } from '../services'
 
 export const REDUCER = 'GOODS'
