--- src/service.js
@@ -1,3 +1,0 @@
-const API = 'http://localhost:4000/'
-
-export { API }
