+++ src/services.js
@@ -1,0 +1,4 @@
+const API = 'http://localhost:4000/'
+const PATH_SRC = './Товары в каталоге Onliner.by_files/'
+
+export { API, PATH_SRC }
